backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Network testing libraries
scapy==2.5.0
ping3==4.0.4
speedtest-cli==2.1.3
dnspython==2.4.2

# AI and Google Gemini
google-generativeai==0.3.2

# Database
supabase==2.9.0
postgrest==0.17.0
websockets==13.0

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
httpx==0.26.0
email-validator==2.1.0

# Logging and monitoring
python-json-logger==2.0.7

# Testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
